feat: setup routes
